requirements.txt
certifi==2020.12.5
Django==2.2.5
djangorestframework==3.12.3
elasticsearch==7.12.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
urllib3==1.26.4